refactor: extract board metrics module
Same forward-reference pattern as data-readers.mjs: readVerdicts,
readPlanCosts, readQAStats, readSecStats still live in server.mjs (moving
to lib/verdicts.mjs next); metrics.mjs imports them circularly from
../server.mjs, safe because calls only happen inside function bodies.
